Update create_chat_tables.php
A column ('data') to store details about each message was added. For instance, if our message type is an image, we can add the image link, or if it is a link we can store the link and attachment details in it. The conversation table already has this column, I think there is a need for this in the message table too.